Don't suggest fully qualified path name
CIFS shares are not added by specifying the path on the host machine.

Instead they are specified by giving the "share name" which on Windows and Mac are likely just the end folders. On Linux this is specified in the samba config.

THIS CHANGE IS INCOMPLETE. The code internally still says path everywhere, which will confuse developers who touch this feature next. PLEASE CHANGE THIS. I have refrained from doing so in this commit because I am not the maintainer of this code and am not entirely sure what the scope of changes I need to make is. I also don't have a quick way to test it so I don't want to make speculative changes since I could break something.